refactor(services): collapse play_history_service + bridge_commands
Two thin service modules absorbed into the larger services that own
their domains. Keeps activity_log.py (4 cross-cutting callers — genuine
utility).

- services/play_history_service.py (57 LOC, 2 callers) → merged into
  services/now_playing.py. Functions get_next_play_order,
  archive_to_history, get_play_history now live adjacent to their
  only-caller domain.

- services/bridge_commands.py (56 LOC, 1 production caller) → merged
  into services/bridge_integration.py. Thread-safe in-memory command
  queue functions queue_command, poll_commands, clear_all.

Notes:
- bridge_integration.py imports archive_to_history *lazily* inside the
  two functions that need it, to avoid a circular import with
  now_playing.py's bottom-of-module re-exports (legacy shim).
- tests/test_bridge_commands.py updated: imports + @patch paths
  retargeted to app.services.bridge_integration.
- Schemas module schemas/bridge_commands.py is unchanged (it was
  already a distinct concern).

Test suite: 151 targeted tests green.
